many additions to settings and catalog pages, some tweaks/fixes in REST resources - jsslab

+++ api/src/main/java/org/openmrs/module/jsslab/db/LabSpecimenTemplateDAO.java
@@ -84,7 +84,7 @@ public interface LabSpecimenTemplateDAO {
 	 * @see org.openmrs.api.LabCatalogService#getCountOfLabSpecimenTemplates(String,
 	 *      Boolean)
 	 */
-	public Integer getCountOfLabSpecimenTemplates(String search,
+	public Long getCountOfLabSpecimenTemplates(String search,
 			Boolean includeRetired);
 
 }
